+++ modules/http-server/src/handler/mod.rs
@@ -1,3 +1,4 @@
+mod file_descriptor;
 mod file_pipeline;
 mod observability;
 #[cfg(any(test, feature = "fuzz"))]